ch(middleware): add middleware
