+++ server/routes/index.js
@@ -2,6 +2,8 @@ import express from 'express';
 
 import ParcelController from '../controllers/parcel';
 import UserController from '../controllers/user';
+import validUser from '../middleware/validateuser';
+import validParcel from '../middleware/validateparcel';
 
 const router = express.Router();
 
@@ -10,7 +12,7 @@ router.get('/parcels', ParcelController.getAllParcels);
 
 router.get('/parcels/:id', ParcelController.getSpecificParcel);
 
-router.post('/parcels', ParcelController.createParcel);
+router.post('/parcels', validParcel, ParcelController.createParcel);
 
 router.put('/parcels/:id/cancel', ParcelController.cancelParcel);
 
@@ -25,7 +27,7 @@ router.get('/users/:id', UserController.getSpecificUser);
 
 router.get('/users/:id/parcels', UserController.getUserParcels);
 
-router.post('/users', UserController.addUserAccount);
+router.post('/users', validUser, UserController.addUserAccount);
 
 router.put('/parcels/:id/deliver', UserController.changeStatus);
